+++ src/components/EditCompany/EditCompany.js
@@ -88,7 +88,7 @@ const EditCompany = () => {
         .then((response) => {
           setCompany(response.data);
           setSubmitted(true);
-          console.log(response.data);
+          // console.log(response.data);
         })
         .catch((e) => {
           console.error(e);