+++ src/components/InsertCompany/InsertCompany.js
@@ -67,7 +67,7 @@ const InsertCompany = () => {
         .then((response) => {
           setCompany(response.data);
           setSubmitted(true);
-          console.log(response.data);
+          // console.log(response.data);
         })
         .catch((e) => {
           console.error(e);